put a second conditon in RoadRunner.__getattr__ such that when item is 'this' then raise an attribute error. This breaks an infinite recursion that has been there for years and only exposed because a dependency of a dependency (x4) increased the recursion limit beyond what Python thinks is acceptible

+++ source/rrRoadRunner.cpp
@@ -1191,7 +1191,7 @@ void RoadRunner::reset()
 
     //reset(opt1);
     uint opt2 = rr::SelectionRecord::DEPENDENT_INITIAL_GLOBAL_PARAMETER;
-    reset(opt2 | opt1);
+    reset((int)opt2 | opt1);
 }
 
 void RoadRunner::reset(int options)
